+++ packages/api/endpoints/iceberg-pdrs.js
@@ -7,12 +7,12 @@
 
 const router = require('express-promise-router')();
 const Logger = require('@cumulus/logger');
-const { PdrSearch } = require('@cumulus/db');
+const { PdrIcebergSearch } = require('@cumulus/db/duckdb');
 
 const log = new Logger({ sender: '@cumulus/api/iceberg-pdrs' });
 
 /**
- * List and search pdrs
+ * List and search PDRs
  *
  * @param {Object} req - express request object
  * @param {Object} res - express response object
@@ -20,9 +20,20 @@ const log = new Logger({ sender: '@cumulus/api/iceberg-pdrs' });
  */
 async function list(req, res) {
   log.debug(`list query ${JSON.stringify(req.query)}`);
-  const search = new PdrSearch({ queryStringParameters: req.query });
-  const response = await search.query();
-  return res.send(response);
+
+  try {
+    const response = await new PdrIcebergSearch({ queryStringParameters: req.query }).query();
+    return res.send(response);
+  } catch (error) {
+    log.error('PdrIcebergSearch Query Failed', error);
+    if (res.boom) {
+      return res.boom.badImplementation('Error querying S3/Iceberg data');
+    }
+    return res.status(500).send({
+      error: 'Internal Server Error',
+      message: 'Error querying S3/Iceberg data',
+    });
+  }
 }
 
 // Only expose the list endpoint for Iceberg API